fix: :bug: Inject firebase authentication configuration as dependency to AuthGate.
