+++ src/gateway-layer/firebase-module/firebaseConfig.ts
@@ -1,4 +1,5 @@
 import { initializeApp } from "firebase/app";
+import { getAuth } from "firebase/auth";
 import { getFirestore } from "firebase/firestore";
 
 const firebaseConfig = {
@@ -12,4 +13,5 @@ const firebaseConfig = {
 
 // Initialize Firebase
 const firebase = initializeApp(firebaseConfig);
+export const firebaseAuth = getAuth(firebase);
 export const firebaseDb = getFirestore(firebase);
